+++ lib/src/kdbx_dao.dart
@@ -1,5 +1,4 @@
 import 'package:clock/clock.dart';
-import 'package:kdbx/src/kdbx_deleted_object.dart';
 import 'package:kdbx/src/kdbx_entry.dart';
 import 'package:kdbx/src/kdbx_file.dart';
 import 'package:kdbx/src/kdbx_group.dart';