+++ test/kdbx_test.dart
@@ -5,7 +5,6 @@ import 'dart:typed_data';
 import 'package:kdbx/kdbx.dart';
 import 'package:kdbx/src/crypto/protected_salt_generator.dart';
 import 'package:logging/logging.dart';
-import 'package:logging_appenders/logging_appenders.dart';
 import 'package:synchronized/synchronized.dart';
 import 'package:test/test.dart';